First iteration HeatMap funcional

+++ P3 - Dades/Assets/P3 - Dades/Scripts/Heatmap/TileMapStruct.cs
@@ -19,5 +19,5 @@ public struct TileMap
 public struct TileStruct
 {
     public GameObject tileGO;
-    public int heat;
+    public float heat;
 }
